+++ CHANGELOG.md
@@ -98,3 +98,4 @@
 * Rename `composition` to `unwrap` and `Outline` to `Primary` in `Composition` typeclass
 * Rename `equip` to `wrap` and `Layout` to `Schema` in `Transformer` typeclass
 * Make `Composition` a superclass for `Transformer` typeclass
+* Move `Composition` and `Transformer` typeclasses to `Pattern` submodule